fix compiler warnings

+++ src/HSDConfigFile.cpp
@@ -11,7 +11,7 @@ HSDConfigFile::~HSDConfigFile()
 {
 }
 
-bool HSDConfigFile::read(char* buffer, int bufSize)
+bool HSDConfigFile::read(char* buffer, size_t bufSize)
 {
   bool success = false;
